// src/app/models/Fournisseur.ts
export class Fournisseur {
    idFournisseur: number;
    code: string;
    libelle: string;
    image_fournisseur: string;
    date_fin_contrat: string = new Date().toDateString();
    likeFournisseur: number;
    email_fournisseur: string;



  /*  constructor(idFournisseur,code,libelle,image_fournisseur,date_fin_contrat,likeFournisseur)
    {
this.idFournisseur=idFournisseur;
this.code=code;
this.libelle=libelle;
this.image_fournisseur=image_fournisseur;
this.date_fin_contrat=date_fin_contrat;
this.likeFournisseur=likeFournisseur;

    }*/


    
  }
  